requirements.txt
pandas
pywhatkit
faiss-cpu
transformers
openpyxl
